+++ installers/dmg_creator/.gitignore
@@ -1,0 +1,1 @@
+*.dmg
