+++ lib/data/requests/requests.dart
@@ -10,3 +10,5 @@ export 'updatepassword_request.dart';
 export 'updateprofile_request.dart';
 export 'upload_request.dart';
 export 'users_request.dart';
+export 'chats_request.dart';
+export 'newproject_request.dart';
