+++ lib/data/responses/options_response.dart
@@ -1,6 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import '../../domain/models/options.dart';
+import '../../domain/models/form_options.dart';
 
 part 'options_response.freezed.dart';
 part 'options_response.g.dart';
@@ -8,7 +8,7 @@ part 'options_response.g.dart';
 @freezed
 class OptionsResponse with _$OptionsResponse {
   factory OptionsResponse({
-    required Options data,
+    required FormOptions data,
   }) = _OptionsResponse;
 
   factory OptionsResponse.fromJson(Map<String, dynamic> json) =>
