+++ lib/data/responses/responses.dart
@@ -7,6 +7,7 @@ export 'messages_response.dart';
 export 'newcomment_response.dart';
 export 'notifications_response.dart';
 export 'offices_response.dart';
+export 'options_response.dart';
 export 'pipsstatus_response.dart';
 export 'project_response.dart';
 export 'projects_response.dart';
@@ -18,3 +19,7 @@ export 'updateprofile_response.dart';
 export 'updateproject_response.dart';
 export 'upload_response.dart';
 export 'users_response.dart';
+export 'createproject_response.dart';
+export 'patchproject_response.dart';
+export 'updatefinancialaccomplishment_response.dart';
+export 'uploadattachment_response.dart';
