+++ lib/domain/models/form_options.freezed.dart
@@ -1,0 +1,1394 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'form_options.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
+FormOptions _$FormOptionsFromJson(Map<String, dynamic> json) {
+  return _FormOptions.fromJson(json);
+}
+
+/// @nodoc
+mixin _$FormOptions {
+  @JsonKey(name: "agenda")
+  List<Option>? get agenda => throw _privateConstructorUsedError;
+  @JsonKey(name: "approval_levels")
+  List<Option>? get approvalLevels => throw _privateConstructorUsedError;
+  @JsonKey(name: "bases")
+  List<Option>? get bases => throw _privateConstructorUsedError;
+  @JsonKey(name: "categories")
+  List<Option>? get categories => throw _privateConstructorUsedError;
+  @JsonKey(name: "cip_types")
+  List<Option>? get cipTypes =>
+      throw _privateConstructorUsedError; // @JsonKey(name: "commodities") List<Option>? commodities,
+  @JsonKey(name: "fs_statuses")
+  List<Option>? get fsStatuses => throw _privateConstructorUsedError;
+  @JsonKey(name: "funding_institutions")
+  List<Option>? get fundingInstitutions => throw _privateConstructorUsedError;
+  @JsonKey(name: "funding_sources")
+  List<Option>? get fundingSources => throw _privateConstructorUsedError;
+  @JsonKey(name: "gads")
+  List<Option>? get gads => throw _privateConstructorUsedError;
+  @JsonKey(name: "implementation_modes")
+  List<Option>? get implementationModes => throw _privateConstructorUsedError;
+  @JsonKey(name: "infrastructure_sectors")
+  List<Option>? get infrastructureSectors =>
+      throw _privateConstructorUsedError; // @JsonKey(name: "locations") List<Option>? locations,
+  @JsonKey(name: "nafmip_outputs")
+  List<Option>? get nafmipOutputs => throw _privateConstructorUsedError;
+  @JsonKey(name: "offices")
+  List<Option>? get offices => throw _privateConstructorUsedError;
+  @JsonKey(name: "operating_units")
+  List<Option>? get operatingUnits => throw _privateConstructorUsedError;
+  @JsonKey(name: "pdp_chapters")
+  List<Option>? get pdpChapters =>
+      throw _privateConstructorUsedError; // @JsonKey(name: "pdp_strategies") List<Option>? pdpStrategies,
+  @JsonKey(name: "pipol_statuses")
+  List<Option>? get pipolStatuses => throw _privateConstructorUsedError;
+  @JsonKey(name: "pips_statuses")
+  List<Option>? get pipsStatuses => throw _privateConstructorUsedError;
+  @JsonKey(name: "preparation_documents")
+  List<Option>? get preparationDocuments => throw _privateConstructorUsedError;
+  @JsonKey(name: "prerequisites")
+  List<Option>? get prerequisites => throw _privateConstructorUsedError;
+  @JsonKey(name: "prexcs")
+  List<Option>? get prexcs => throw _privateConstructorUsedError;
+  @JsonKey(name: "programs")
+  List<Option>? get programs => throw _privateConstructorUsedError;
+  @JsonKey(name: "project_statuses")
+  List<Option>? get projectStatuses => throw _privateConstructorUsedError;
+  @JsonKey(name: "provinces")
+  List<Option>? get provinces => throw _privateConstructorUsedError;
+  @JsonKey(name: "readiness_levels")
+  List<Option>? get readinessLevels => throw _privateConstructorUsedError;
+  @JsonKey(name: "regions")
+  List<Option>? get regions => throw _privateConstructorUsedError;
+  @JsonKey(name: "roles")
+  List<Option>? get roles => throw _privateConstructorUsedError;
+  @JsonKey(name: "sdgs")
+  List<Option>? get sdgs => throw _privateConstructorUsedError;
+  @JsonKey(name: "spatial_coverages")
+  List<Option>? get spatialCoverages => throw _privateConstructorUsedError;
+  @JsonKey(name: "types")
+  List<Option>? get types => throw _privateConstructorUsedError;
+  @JsonKey(name: "typologies")
+  List<Option>? get typologies =>
+      throw _privateConstructorUsedError; // @JsonKey(name: "vc_segments") List<Option>? vcSegments,
+  @JsonKey(name: "years")
+  List<Option>? get years => throw _privateConstructorUsedError;
+  @JsonKey(name: "ou_types")
+  List<Option>? get ouTypes => throw _privateConstructorUsedError;
+  @JsonKey(name: "attachment_types")
+  List<Option>? get attachmentTypes => throw _privateConstructorUsedError;
+
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $FormOptionsCopyWith<FormOptions> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $FormOptionsCopyWith<$Res> {
+  factory $FormOptionsCopyWith(
+          FormOptions value, $Res Function(FormOptions) then) =
+      _$FormOptionsCopyWithImpl<$Res, FormOptions>;
+  @useResult
+  $Res call(
+      {@JsonKey(name: "agenda")
+          List<Option>? agenda,
+      @JsonKey(name: "approval_levels")
+          List<Option>? approvalLevels,
+      @JsonKey(name: "bases")
+          List<Option>? bases,
+      @JsonKey(name: "categories")
+          List<Option>? categories,
+      @JsonKey(name: "cip_types")
+          List<Option>? cipTypes,
+      @JsonKey(name: "fs_statuses")
+          List<Option>? fsStatuses,
+      @JsonKey(name: "funding_institutions")
+          List<Option>? fundingInstitutions,
+      @JsonKey(name: "funding_sources")
+          List<Option>? fundingSources,
+      @JsonKey(name: "gads")
+          List<Option>? gads,
+      @JsonKey(name: "implementation_modes")
+          List<Option>? implementationModes,
+      @JsonKey(name: "infrastructure_sectors")
+          List<Option>? infrastructureSectors,
+      @JsonKey(name: "nafmip_outputs")
+          List<Option>? nafmipOutputs,
+      @JsonKey(name: "offices")
+          List<Option>? offices,
+      @JsonKey(name: "operating_units")
+          List<Option>? operatingUnits,
+      @JsonKey(name: "pdp_chapters")
+          List<Option>? pdpChapters,
+      @JsonKey(name: "pipol_statuses")
+          List<Option>? pipolStatuses,
+      @JsonKey(name: "pips_statuses")
+          List<Option>? pipsStatuses,
+      @JsonKey(name: "preparation_documents")
+          List<Option>? preparationDocuments,
+      @JsonKey(name: "prerequisites")
+          List<Option>? prerequisites,
+      @JsonKey(name: "prexcs")
+          List<Option>? prexcs,
+      @JsonKey(name: "programs")
+          List<Option>? programs,
+      @JsonKey(name: "project_statuses")
+          List<Option>? projectStatuses,
+      @JsonKey(name: "provinces")
+          List<Option>? provinces,
+      @JsonKey(name: "readiness_levels")
+          List<Option>? readinessLevels,
+      @JsonKey(name: "regions")
+          List<Option>? regions,
+      @JsonKey(name: "roles")
+          List<Option>? roles,
+      @JsonKey(name: "sdgs")
+          List<Option>? sdgs,
+      @JsonKey(name: "spatial_coverages")
+          List<Option>? spatialCoverages,
+      @JsonKey(name: "types")
+          List<Option>? types,
+      @JsonKey(name: "typologies")
+          List<Option>? typologies,
+      @JsonKey(name: "years")
+          List<Option>? years,
+      @JsonKey(name: "ou_types")
+          List<Option>? ouTypes,
+      @JsonKey(name: "attachment_types")
+          List<Option>? attachmentTypes});
+}
+
+/// @nodoc
+class _$FormOptionsCopyWithImpl<$Res, $Val extends FormOptions>
+    implements $FormOptionsCopyWith<$Res> {
+  _$FormOptionsCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? agenda = freezed,
+    Object? approvalLevels = freezed,
+    Object? bases = freezed,
+    Object? categories = freezed,
+    Object? cipTypes = freezed,
+    Object? fsStatuses = freezed,
+    Object? fundingInstitutions = freezed,
+    Object? fundingSources = freezed,
+    Object? gads = freezed,
+    Object? implementationModes = freezed,
+    Object? infrastructureSectors = freezed,
+    Object? nafmipOutputs = freezed,
+    Object? offices = freezed,
+    Object? operatingUnits = freezed,
+    Object? pdpChapters = freezed,
+    Object? pipolStatuses = freezed,
+    Object? pipsStatuses = freezed,
+    Object? preparationDocuments = freezed,
+    Object? prerequisites = freezed,
+    Object? prexcs = freezed,
+    Object? programs = freezed,
+    Object? projectStatuses = freezed,
+    Object? provinces = freezed,
+    Object? readinessLevels = freezed,
+    Object? regions = freezed,
+    Object? roles = freezed,
+    Object? sdgs = freezed,
+    Object? spatialCoverages = freezed,
+    Object? types = freezed,
+    Object? typologies = freezed,
+    Object? years = freezed,
+    Object? ouTypes = freezed,
+    Object? attachmentTypes = freezed,
+  }) {
+    return _then(_value.copyWith(
+      agenda: freezed == agenda
+          ? _value.agenda
+          : agenda // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      approvalLevels: freezed == approvalLevels
+          ? _value.approvalLevels
+          : approvalLevels // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      bases: freezed == bases
+          ? _value.bases
+          : bases // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      categories: freezed == categories
+          ? _value.categories
+          : categories // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      cipTypes: freezed == cipTypes
+          ? _value.cipTypes
+          : cipTypes // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      fsStatuses: freezed == fsStatuses
+          ? _value.fsStatuses
+          : fsStatuses // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      fundingInstitutions: freezed == fundingInstitutions
+          ? _value.fundingInstitutions
+          : fundingInstitutions // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      fundingSources: freezed == fundingSources
+          ? _value.fundingSources
+          : fundingSources // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      gads: freezed == gads
+          ? _value.gads
+          : gads // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      implementationModes: freezed == implementationModes
+          ? _value.implementationModes
+          : implementationModes // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      infrastructureSectors: freezed == infrastructureSectors
+          ? _value.infrastructureSectors
+          : infrastructureSectors // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      nafmipOutputs: freezed == nafmipOutputs
+          ? _value.nafmipOutputs
+          : nafmipOutputs // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      offices: freezed == offices
+          ? _value.offices
+          : offices // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      operatingUnits: freezed == operatingUnits
+          ? _value.operatingUnits
+          : operatingUnits // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      pdpChapters: freezed == pdpChapters
+          ? _value.pdpChapters
+          : pdpChapters // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      pipolStatuses: freezed == pipolStatuses
+          ? _value.pipolStatuses
+          : pipolStatuses // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      pipsStatuses: freezed == pipsStatuses
+          ? _value.pipsStatuses
+          : pipsStatuses // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      preparationDocuments: freezed == preparationDocuments
+          ? _value.preparationDocuments
+          : preparationDocuments // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      prerequisites: freezed == prerequisites
+          ? _value.prerequisites
+          : prerequisites // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      prexcs: freezed == prexcs
+          ? _value.prexcs
+          : prexcs // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      programs: freezed == programs
+          ? _value.programs
+          : programs // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      projectStatuses: freezed == projectStatuses
+          ? _value.projectStatuses
+          : projectStatuses // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      provinces: freezed == provinces
+          ? _value.provinces
+          : provinces // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      readinessLevels: freezed == readinessLevels
+          ? _value.readinessLevels
+          : readinessLevels // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      regions: freezed == regions
+          ? _value.regions
+          : regions // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      roles: freezed == roles
+          ? _value.roles
+          : roles // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      sdgs: freezed == sdgs
+          ? _value.sdgs
+          : sdgs // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      spatialCoverages: freezed == spatialCoverages
+          ? _value.spatialCoverages
+          : spatialCoverages // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      types: freezed == types
+          ? _value.types
+          : types // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      typologies: freezed == typologies
+          ? _value.typologies
+          : typologies // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      years: freezed == years
+          ? _value.years
+          : years // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      ouTypes: freezed == ouTypes
+          ? _value.ouTypes
+          : ouTypes // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      attachmentTypes: freezed == attachmentTypes
+          ? _value.attachmentTypes
+          : attachmentTypes // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+    ) as $Val);
+  }
+}
+
+/// @nodoc
+abstract class _$$_FormOptionsCopyWith<$Res>
+    implements $FormOptionsCopyWith<$Res> {
+  factory _$$_FormOptionsCopyWith(
+          _$_FormOptions value, $Res Function(_$_FormOptions) then) =
+      __$$_FormOptionsCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call(
+      {@JsonKey(name: "agenda")
+          List<Option>? agenda,
+      @JsonKey(name: "approval_levels")
+          List<Option>? approvalLevels,
+      @JsonKey(name: "bases")
+          List<Option>? bases,
+      @JsonKey(name: "categories")
+          List<Option>? categories,
+      @JsonKey(name: "cip_types")
+          List<Option>? cipTypes,
+      @JsonKey(name: "fs_statuses")
+          List<Option>? fsStatuses,
+      @JsonKey(name: "funding_institutions")
+          List<Option>? fundingInstitutions,
+      @JsonKey(name: "funding_sources")
+          List<Option>? fundingSources,
+      @JsonKey(name: "gads")
+          List<Option>? gads,
+      @JsonKey(name: "implementation_modes")
+          List<Option>? implementationModes,
+      @JsonKey(name: "infrastructure_sectors")
+          List<Option>? infrastructureSectors,
+      @JsonKey(name: "nafmip_outputs")
+          List<Option>? nafmipOutputs,
+      @JsonKey(name: "offices")
+          List<Option>? offices,
+      @JsonKey(name: "operating_units")
+          List<Option>? operatingUnits,
+      @JsonKey(name: "pdp_chapters")
+          List<Option>? pdpChapters,
+      @JsonKey(name: "pipol_statuses")
+          List<Option>? pipolStatuses,
+      @JsonKey(name: "pips_statuses")
+          List<Option>? pipsStatuses,
+      @JsonKey(name: "preparation_documents")
+          List<Option>? preparationDocuments,
+      @JsonKey(name: "prerequisites")
+          List<Option>? prerequisites,
+      @JsonKey(name: "prexcs")
+          List<Option>? prexcs,
+      @JsonKey(name: "programs")
+          List<Option>? programs,
+      @JsonKey(name: "project_statuses")
+          List<Option>? projectStatuses,
+      @JsonKey(name: "provinces")
+          List<Option>? provinces,
+      @JsonKey(name: "readiness_levels")
+          List<Option>? readinessLevels,
+      @JsonKey(name: "regions")
+          List<Option>? regions,
+      @JsonKey(name: "roles")
+          List<Option>? roles,
+      @JsonKey(name: "sdgs")
+          List<Option>? sdgs,
+      @JsonKey(name: "spatial_coverages")
+          List<Option>? spatialCoverages,
+      @JsonKey(name: "types")
+          List<Option>? types,
+      @JsonKey(name: "typologies")
+          List<Option>? typologies,
+      @JsonKey(name: "years")
+          List<Option>? years,
+      @JsonKey(name: "ou_types")
+          List<Option>? ouTypes,
+      @JsonKey(name: "attachment_types")
+          List<Option>? attachmentTypes});
+}
+
+/// @nodoc
+class __$$_FormOptionsCopyWithImpl<$Res>
+    extends _$FormOptionsCopyWithImpl<$Res, _$_FormOptions>
+    implements _$$_FormOptionsCopyWith<$Res> {
+  __$$_FormOptionsCopyWithImpl(
+      _$_FormOptions _value, $Res Function(_$_FormOptions) _then)
+      : super(_value, _then);
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? agenda = freezed,
+    Object? approvalLevels = freezed,
+    Object? bases = freezed,
+    Object? categories = freezed,
+    Object? cipTypes = freezed,
+    Object? fsStatuses = freezed,
+    Object? fundingInstitutions = freezed,
+    Object? fundingSources = freezed,
+    Object? gads = freezed,
+    Object? implementationModes = freezed,
+    Object? infrastructureSectors = freezed,
+    Object? nafmipOutputs = freezed,
+    Object? offices = freezed,
+    Object? operatingUnits = freezed,
+    Object? pdpChapters = freezed,
+    Object? pipolStatuses = freezed,
+    Object? pipsStatuses = freezed,
+    Object? preparationDocuments = freezed,
+    Object? prerequisites = freezed,
+    Object? prexcs = freezed,
+    Object? programs = freezed,
+    Object? projectStatuses = freezed,
+    Object? provinces = freezed,
+    Object? readinessLevels = freezed,
+    Object? regions = freezed,
+    Object? roles = freezed,
+    Object? sdgs = freezed,
+    Object? spatialCoverages = freezed,
+    Object? types = freezed,
+    Object? typologies = freezed,
+    Object? years = freezed,
+    Object? ouTypes = freezed,
+    Object? attachmentTypes = freezed,
+  }) {
+    return _then(_$_FormOptions(
+      agenda: freezed == agenda
+          ? _value._agenda
+          : agenda // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      approvalLevels: freezed == approvalLevels
+          ? _value._approvalLevels
+          : approvalLevels // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      bases: freezed == bases
+          ? _value._bases
+          : bases // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      categories: freezed == categories
+          ? _value._categories
+          : categories // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      cipTypes: freezed == cipTypes
+          ? _value._cipTypes
+          : cipTypes // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      fsStatuses: freezed == fsStatuses
+          ? _value._fsStatuses
+          : fsStatuses // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      fundingInstitutions: freezed == fundingInstitutions
+          ? _value._fundingInstitutions
+          : fundingInstitutions // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      fundingSources: freezed == fundingSources
+          ? _value._fundingSources
+          : fundingSources // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      gads: freezed == gads
+          ? _value._gads
+          : gads // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      implementationModes: freezed == implementationModes
+          ? _value._implementationModes
+          : implementationModes // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      infrastructureSectors: freezed == infrastructureSectors
+          ? _value._infrastructureSectors
+          : infrastructureSectors // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      nafmipOutputs: freezed == nafmipOutputs
+          ? _value._nafmipOutputs
+          : nafmipOutputs // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      offices: freezed == offices
+          ? _value._offices
+          : offices // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      operatingUnits: freezed == operatingUnits
+          ? _value._operatingUnits
+          : operatingUnits // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      pdpChapters: freezed == pdpChapters
+          ? _value._pdpChapters
+          : pdpChapters // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      pipolStatuses: freezed == pipolStatuses
+          ? _value._pipolStatuses
+          : pipolStatuses // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      pipsStatuses: freezed == pipsStatuses
+          ? _value._pipsStatuses
+          : pipsStatuses // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      preparationDocuments: freezed == preparationDocuments
+          ? _value._preparationDocuments
+          : preparationDocuments // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      prerequisites: freezed == prerequisites
+          ? _value._prerequisites
+          : prerequisites // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      prexcs: freezed == prexcs
+          ? _value._prexcs
+          : prexcs // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      programs: freezed == programs
+          ? _value._programs
+          : programs // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      projectStatuses: freezed == projectStatuses
+          ? _value._projectStatuses
+          : projectStatuses // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      provinces: freezed == provinces
+          ? _value._provinces
+          : provinces // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      readinessLevels: freezed == readinessLevels
+          ? _value._readinessLevels
+          : readinessLevels // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      regions: freezed == regions
+          ? _value._regions
+          : regions // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      roles: freezed == roles
+          ? _value._roles
+          : roles // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      sdgs: freezed == sdgs
+          ? _value._sdgs
+          : sdgs // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      spatialCoverages: freezed == spatialCoverages
+          ? _value._spatialCoverages
+          : spatialCoverages // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      types: freezed == types
+          ? _value._types
+          : types // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      typologies: freezed == typologies
+          ? _value._typologies
+          : typologies // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      years: freezed == years
+          ? _value._years
+          : years // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      ouTypes: freezed == ouTypes
+          ? _value._ouTypes
+          : ouTypes // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+      attachmentTypes: freezed == attachmentTypes
+          ? _value._attachmentTypes
+          : attachmentTypes // ignore: cast_nullable_to_non_nullable
+              as List<Option>?,
+    ));
+  }
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$_FormOptions implements _FormOptions {
+  _$_FormOptions(
+      {@JsonKey(name: "agenda")
+          final List<Option>? agenda,
+      @JsonKey(name: "approval_levels")
+          final List<Option>? approvalLevels,
+      @JsonKey(name: "bases")
+          final List<Option>? bases,
+      @JsonKey(name: "categories")
+          final List<Option>? categories,
+      @JsonKey(name: "cip_types")
+          final List<Option>? cipTypes,
+      @JsonKey(name: "fs_statuses")
+          final List<Option>? fsStatuses,
+      @JsonKey(name: "funding_institutions")
+          final List<Option>? fundingInstitutions,
+      @JsonKey(name: "funding_sources")
+          final List<Option>? fundingSources,
+      @JsonKey(name: "gads")
+          final List<Option>? gads,
+      @JsonKey(name: "implementation_modes")
+          final List<Option>? implementationModes,
+      @JsonKey(name: "infrastructure_sectors")
+          final List<Option>? infrastructureSectors,
+      @JsonKey(name: "nafmip_outputs")
+          final List<Option>? nafmipOutputs,
+      @JsonKey(name: "offices")
+          final List<Option>? offices,
+      @JsonKey(name: "operating_units")
+          final List<Option>? operatingUnits,
+      @JsonKey(name: "pdp_chapters")
+          final List<Option>? pdpChapters,
+      @JsonKey(name: "pipol_statuses")
+          final List<Option>? pipolStatuses,
+      @JsonKey(name: "pips_statuses")
+          final List<Option>? pipsStatuses,
+      @JsonKey(name: "preparation_documents")
+          final List<Option>? preparationDocuments,
+      @JsonKey(name: "prerequisites")
+          final List<Option>? prerequisites,
+      @JsonKey(name: "prexcs")
+          final List<Option>? prexcs,
+      @JsonKey(name: "programs")
+          final List<Option>? programs,
+      @JsonKey(name: "project_statuses")
+          final List<Option>? projectStatuses,
+      @JsonKey(name: "provinces")
+          final List<Option>? provinces,
+      @JsonKey(name: "readiness_levels")
+          final List<Option>? readinessLevels,
+      @JsonKey(name: "regions")
+          final List<Option>? regions,
+      @JsonKey(name: "roles")
+          final List<Option>? roles,
+      @JsonKey(name: "sdgs")
+          final List<Option>? sdgs,
+      @JsonKey(name: "spatial_coverages")
+          final List<Option>? spatialCoverages,
+      @JsonKey(name: "types")
+          final List<Option>? types,
+      @JsonKey(name: "typologies")
+          final List<Option>? typologies,
+      @JsonKey(name: "years")
+          final List<Option>? years,
+      @JsonKey(name: "ou_types")
+          final List<Option>? ouTypes,
+      @JsonKey(name: "attachment_types")
+          final List<Option>? attachmentTypes})
+      : _agenda = agenda,
+        _approvalLevels = approvalLevels,
+        _bases = bases,
+        _categories = categories,
+        _cipTypes = cipTypes,
+        _fsStatuses = fsStatuses,
+        _fundingInstitutions = fundingInstitutions,
+        _fundingSources = fundingSources,
+        _gads = gads,
+        _implementationModes = implementationModes,
+        _infrastructureSectors = infrastructureSectors,
+        _nafmipOutputs = nafmipOutputs,
+        _offices = offices,
+        _operatingUnits = operatingUnits,
+        _pdpChapters = pdpChapters,
+        _pipolStatuses = pipolStatuses,
+        _pipsStatuses = pipsStatuses,
+        _preparationDocuments = preparationDocuments,
+        _prerequisites = prerequisites,
+        _prexcs = prexcs,
+        _programs = programs,
+        _projectStatuses = projectStatuses,
+        _provinces = provinces,
+        _readinessLevels = readinessLevels,
+        _regions = regions,
+        _roles = roles,
+        _sdgs = sdgs,
+        _spatialCoverages = spatialCoverages,
+        _types = types,
+        _typologies = typologies,
+        _years = years,
+        _ouTypes = ouTypes,
+        _attachmentTypes = attachmentTypes;
+
+  factory _$_FormOptions.fromJson(Map<String, dynamic> json) =>
+      _$$_FormOptionsFromJson(json);
+
+  final List<Option>? _agenda;
+  @override
+  @JsonKey(name: "agenda")
+  List<Option>? get agenda {
+    final value = _agenda;
+    if (value == null) return null;
+    if (_agenda is EqualUnmodifiableListView) return _agenda;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _approvalLevels;
+  @override
+  @JsonKey(name: "approval_levels")
+  List<Option>? get approvalLevels {
+    final value = _approvalLevels;
+    if (value == null) return null;
+    if (_approvalLevels is EqualUnmodifiableListView) return _approvalLevels;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _bases;
+  @override
+  @JsonKey(name: "bases")
+  List<Option>? get bases {
+    final value = _bases;
+    if (value == null) return null;
+    if (_bases is EqualUnmodifiableListView) return _bases;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _categories;
+  @override
+  @JsonKey(name: "categories")
+  List<Option>? get categories {
+    final value = _categories;
+    if (value == null) return null;
+    if (_categories is EqualUnmodifiableListView) return _categories;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _cipTypes;
+  @override
+  @JsonKey(name: "cip_types")
+  List<Option>? get cipTypes {
+    final value = _cipTypes;
+    if (value == null) return null;
+    if (_cipTypes is EqualUnmodifiableListView) return _cipTypes;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+// @JsonKey(name: "commodities") List<Option>? commodities,
+  final List<Option>? _fsStatuses;
+// @JsonKey(name: "commodities") List<Option>? commodities,
+  @override
+  @JsonKey(name: "fs_statuses")
+  List<Option>? get fsStatuses {
+    final value = _fsStatuses;
+    if (value == null) return null;
+    if (_fsStatuses is EqualUnmodifiableListView) return _fsStatuses;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _fundingInstitutions;
+  @override
+  @JsonKey(name: "funding_institutions")
+  List<Option>? get fundingInstitutions {
+    final value = _fundingInstitutions;
+    if (value == null) return null;
+    if (_fundingInstitutions is EqualUnmodifiableListView)
+      return _fundingInstitutions;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _fundingSources;
+  @override
+  @JsonKey(name: "funding_sources")
+  List<Option>? get fundingSources {
+    final value = _fundingSources;
+    if (value == null) return null;
+    if (_fundingSources is EqualUnmodifiableListView) return _fundingSources;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _gads;
+  @override
+  @JsonKey(name: "gads")
+  List<Option>? get gads {
+    final value = _gads;
+    if (value == null) return null;
+    if (_gads is EqualUnmodifiableListView) return _gads;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _implementationModes;
+  @override
+  @JsonKey(name: "implementation_modes")
+  List<Option>? get implementationModes {
+    final value = _implementationModes;
+    if (value == null) return null;
+    if (_implementationModes is EqualUnmodifiableListView)
+      return _implementationModes;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _infrastructureSectors;
+  @override
+  @JsonKey(name: "infrastructure_sectors")
+  List<Option>? get infrastructureSectors {
+    final value = _infrastructureSectors;
+    if (value == null) return null;
+    if (_infrastructureSectors is EqualUnmodifiableListView)
+      return _infrastructureSectors;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+// @JsonKey(name: "locations") List<Option>? locations,
+  final List<Option>? _nafmipOutputs;
+// @JsonKey(name: "locations") List<Option>? locations,
+  @override
+  @JsonKey(name: "nafmip_outputs")
+  List<Option>? get nafmipOutputs {
+    final value = _nafmipOutputs;
+    if (value == null) return null;
+    if (_nafmipOutputs is EqualUnmodifiableListView) return _nafmipOutputs;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _offices;
+  @override
+  @JsonKey(name: "offices")
+  List<Option>? get offices {
+    final value = _offices;
+    if (value == null) return null;
+    if (_offices is EqualUnmodifiableListView) return _offices;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _operatingUnits;
+  @override
+  @JsonKey(name: "operating_units")
+  List<Option>? get operatingUnits {
+    final value = _operatingUnits;
+    if (value == null) return null;
+    if (_operatingUnits is EqualUnmodifiableListView) return _operatingUnits;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _pdpChapters;
+  @override
+  @JsonKey(name: "pdp_chapters")
+  List<Option>? get pdpChapters {
+    final value = _pdpChapters;
+    if (value == null) return null;
+    if (_pdpChapters is EqualUnmodifiableListView) return _pdpChapters;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+// @JsonKey(name: "pdp_strategies") List<Option>? pdpStrategies,
+  final List<Option>? _pipolStatuses;
+// @JsonKey(name: "pdp_strategies") List<Option>? pdpStrategies,
+  @override
+  @JsonKey(name: "pipol_statuses")
+  List<Option>? get pipolStatuses {
+    final value = _pipolStatuses;
+    if (value == null) return null;
+    if (_pipolStatuses is EqualUnmodifiableListView) return _pipolStatuses;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _pipsStatuses;
+  @override
+  @JsonKey(name: "pips_statuses")
+  List<Option>? get pipsStatuses {
+    final value = _pipsStatuses;
+    if (value == null) return null;
+    if (_pipsStatuses is EqualUnmodifiableListView) return _pipsStatuses;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _preparationDocuments;
+  @override
+  @JsonKey(name: "preparation_documents")
+  List<Option>? get preparationDocuments {
+    final value = _preparationDocuments;
+    if (value == null) return null;
+    if (_preparationDocuments is EqualUnmodifiableListView)
+      return _preparationDocuments;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _prerequisites;
+  @override
+  @JsonKey(name: "prerequisites")
+  List<Option>? get prerequisites {
+    final value = _prerequisites;
+    if (value == null) return null;
+    if (_prerequisites is EqualUnmodifiableListView) return _prerequisites;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _prexcs;
+  @override
+  @JsonKey(name: "prexcs")
+  List<Option>? get prexcs {
+    final value = _prexcs;
+    if (value == null) return null;
+    if (_prexcs is EqualUnmodifiableListView) return _prexcs;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _programs;
+  @override
+  @JsonKey(name: "programs")
+  List<Option>? get programs {
+    final value = _programs;
+    if (value == null) return null;
+    if (_programs is EqualUnmodifiableListView) return _programs;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _projectStatuses;
+  @override
+  @JsonKey(name: "project_statuses")
+  List<Option>? get projectStatuses {
+    final value = _projectStatuses;
+    if (value == null) return null;
+    if (_projectStatuses is EqualUnmodifiableListView) return _projectStatuses;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _provinces;
+  @override
+  @JsonKey(name: "provinces")
+  List<Option>? get provinces {
+    final value = _provinces;
+    if (value == null) return null;
+    if (_provinces is EqualUnmodifiableListView) return _provinces;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _readinessLevels;
+  @override
+  @JsonKey(name: "readiness_levels")
+  List<Option>? get readinessLevels {
+    final value = _readinessLevels;
+    if (value == null) return null;
+    if (_readinessLevels is EqualUnmodifiableListView) return _readinessLevels;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _regions;
+  @override
+  @JsonKey(name: "regions")
+  List<Option>? get regions {
+    final value = _regions;
+    if (value == null) return null;
+    if (_regions is EqualUnmodifiableListView) return _regions;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _roles;
+  @override
+  @JsonKey(name: "roles")
+  List<Option>? get roles {
+    final value = _roles;
+    if (value == null) return null;
+    if (_roles is EqualUnmodifiableListView) return _roles;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _sdgs;
+  @override
+  @JsonKey(name: "sdgs")
+  List<Option>? get sdgs {
+    final value = _sdgs;
+    if (value == null) return null;
+    if (_sdgs is EqualUnmodifiableListView) return _sdgs;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _spatialCoverages;
+  @override
+  @JsonKey(name: "spatial_coverages")
+  List<Option>? get spatialCoverages {
+    final value = _spatialCoverages;
+    if (value == null) return null;
+    if (_spatialCoverages is EqualUnmodifiableListView)
+      return _spatialCoverages;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _types;
+  @override
+  @JsonKey(name: "types")
+  List<Option>? get types {
+    final value = _types;
+    if (value == null) return null;
+    if (_types is EqualUnmodifiableListView) return _types;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _typologies;
+  @override
+  @JsonKey(name: "typologies")
+  List<Option>? get typologies {
+    final value = _typologies;
+    if (value == null) return null;
+    if (_typologies is EqualUnmodifiableListView) return _typologies;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+// @JsonKey(name: "vc_segments") List<Option>? vcSegments,
+  final List<Option>? _years;
+// @JsonKey(name: "vc_segments") List<Option>? vcSegments,
+  @override
+  @JsonKey(name: "years")
+  List<Option>? get years {
+    final value = _years;
+    if (value == null) return null;
+    if (_years is EqualUnmodifiableListView) return _years;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _ouTypes;
+  @override
+  @JsonKey(name: "ou_types")
+  List<Option>? get ouTypes {
+    final value = _ouTypes;
+    if (value == null) return null;
+    if (_ouTypes is EqualUnmodifiableListView) return _ouTypes;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  final List<Option>? _attachmentTypes;
+  @override
+  @JsonKey(name: "attachment_types")
+  List<Option>? get attachmentTypes {
+    final value = _attachmentTypes;
+    if (value == null) return null;
+    if (_attachmentTypes is EqualUnmodifiableListView) return _attachmentTypes;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  @override
+  String toString() {
+    return 'FormOptions(agenda: $agenda, approvalLevels: $approvalLevels, bases: $bases, categories: $categories, cipTypes: $cipTypes, fsStatuses: $fsStatuses, fundingInstitutions: $fundingInstitutions, fundingSources: $fundingSources, gads: $gads, implementationModes: $implementationModes, infrastructureSectors: $infrastructureSectors, nafmipOutputs: $nafmipOutputs, offices: $offices, operatingUnits: $operatingUnits, pdpChapters: $pdpChapters, pipolStatuses: $pipolStatuses, pipsStatuses: $pipsStatuses, preparationDocuments: $preparationDocuments, prerequisites: $prerequisites, prexcs: $prexcs, programs: $programs, projectStatuses: $projectStatuses, provinces: $provinces, readinessLevels: $readinessLevels, regions: $regions, roles: $roles, sdgs: $sdgs, spatialCoverages: $spatialCoverages, types: $types, typologies: $typologies, years: $years, ouTypes: $ouTypes, attachmentTypes: $attachmentTypes)';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$_FormOptions &&
+            const DeepCollectionEquality().equals(other._agenda, _agenda) &&
+            const DeepCollectionEquality()
+                .equals(other._approvalLevels, _approvalLevels) &&
+            const DeepCollectionEquality().equals(other._bases, _bases) &&
+            const DeepCollectionEquality()
+                .equals(other._categories, _categories) &&
+            const DeepCollectionEquality().equals(other._cipTypes, _cipTypes) &&
+            const DeepCollectionEquality()
+                .equals(other._fsStatuses, _fsStatuses) &&
+            const DeepCollectionEquality()
+                .equals(other._fundingInstitutions, _fundingInstitutions) &&
+            const DeepCollectionEquality()
+                .equals(other._fundingSources, _fundingSources) &&
+            const DeepCollectionEquality().equals(other._gads, _gads) &&
+            const DeepCollectionEquality()
+                .equals(other._implementationModes, _implementationModes) &&
+            const DeepCollectionEquality()
+                .equals(other._infrastructureSectors, _infrastructureSectors) &&
+            const DeepCollectionEquality()
+                .equals(other._nafmipOutputs, _nafmipOutputs) &&
+            const DeepCollectionEquality().equals(other._offices, _offices) &&
+            const DeepCollectionEquality()
+                .equals(other._operatingUnits, _operatingUnits) &&
+            const DeepCollectionEquality()
+                .equals(other._pdpChapters, _pdpChapters) &&
+            const DeepCollectionEquality()
+                .equals(other._pipolStatuses, _pipolStatuses) &&
+            const DeepCollectionEquality()
+                .equals(other._pipsStatuses, _pipsStatuses) &&
+            const DeepCollectionEquality()
+                .equals(other._preparationDocuments, _preparationDocuments) &&
+            const DeepCollectionEquality()
+                .equals(other._prerequisites, _prerequisites) &&
+            const DeepCollectionEquality().equals(other._prexcs, _prexcs) &&
+            const DeepCollectionEquality().equals(other._programs, _programs) &&
+            const DeepCollectionEquality()
+                .equals(other._projectStatuses, _projectStatuses) &&
+            const DeepCollectionEquality()
+                .equals(other._provinces, _provinces) &&
+            const DeepCollectionEquality()
+                .equals(other._readinessLevels, _readinessLevels) &&
+            const DeepCollectionEquality().equals(other._regions, _regions) &&
+            const DeepCollectionEquality().equals(other._roles, _roles) &&
+            const DeepCollectionEquality().equals(other._sdgs, _sdgs) &&
+            const DeepCollectionEquality()
+                .equals(other._spatialCoverages, _spatialCoverages) &&
+            const DeepCollectionEquality().equals(other._types, _types) &&
+            const DeepCollectionEquality()
+                .equals(other._typologies, _typologies) &&
+            const DeepCollectionEquality().equals(other._years, _years) &&
+            const DeepCollectionEquality().equals(other._ouTypes, _ouTypes) &&
+            const DeepCollectionEquality()
+                .equals(other._attachmentTypes, _attachmentTypes));
+  }
+
+  @JsonKey(ignore: true)
+  @override
+  int get hashCode => Object.hashAll([
+        runtimeType,
+        const DeepCollectionEquality().hash(_agenda),
+        const DeepCollectionEquality().hash(_approvalLevels),
+        const DeepCollectionEquality().hash(_bases),
+        const DeepCollectionEquality().hash(_categories),
+        const DeepCollectionEquality().hash(_cipTypes),
+        const DeepCollectionEquality().hash(_fsStatuses),
+        const DeepCollectionEquality().hash(_fundingInstitutions),
+        const DeepCollectionEquality().hash(_fundingSources),
+        const DeepCollectionEquality().hash(_gads),
+        const DeepCollectionEquality().hash(_implementationModes),
+        const DeepCollectionEquality().hash(_infrastructureSectors),
+        const DeepCollectionEquality().hash(_nafmipOutputs),
+        const DeepCollectionEquality().hash(_offices),
+        const DeepCollectionEquality().hash(_operatingUnits),
+        const DeepCollectionEquality().hash(_pdpChapters),
+        const DeepCollectionEquality().hash(_pipolStatuses),
+        const DeepCollectionEquality().hash(_pipsStatuses),
+        const DeepCollectionEquality().hash(_preparationDocuments),
+        const DeepCollectionEquality().hash(_prerequisites),
+        const DeepCollectionEquality().hash(_prexcs),
+        const DeepCollectionEquality().hash(_programs),
+        const DeepCollectionEquality().hash(_projectStatuses),
+        const DeepCollectionEquality().hash(_provinces),
+        const DeepCollectionEquality().hash(_readinessLevels),
+        const DeepCollectionEquality().hash(_regions),
+        const DeepCollectionEquality().hash(_roles),
+        const DeepCollectionEquality().hash(_sdgs),
+        const DeepCollectionEquality().hash(_spatialCoverages),
+        const DeepCollectionEquality().hash(_types),
+        const DeepCollectionEquality().hash(_typologies),
+        const DeepCollectionEquality().hash(_years),
+        const DeepCollectionEquality().hash(_ouTypes),
+        const DeepCollectionEquality().hash(_attachmentTypes)
+      ]);
+
+  @JsonKey(ignore: true)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$_FormOptionsCopyWith<_$_FormOptions> get copyWith =>
+      __$$_FormOptionsCopyWithImpl<_$_FormOptions>(this, _$identity);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$$_FormOptionsToJson(
+      this,
+    );
+  }
+}
+
+abstract class _FormOptions implements FormOptions {
+  factory _FormOptions(
+      {@JsonKey(name: "agenda")
+          final List<Option>? agenda,
+      @JsonKey(name: "approval_levels")
+          final List<Option>? approvalLevels,
+      @JsonKey(name: "bases")
+          final List<Option>? bases,
+      @JsonKey(name: "categories")
+          final List<Option>? categories,
+      @JsonKey(name: "cip_types")
+          final List<Option>? cipTypes,
+      @JsonKey(name: "fs_statuses")
+          final List<Option>? fsStatuses,
+      @JsonKey(name: "funding_institutions")
+          final List<Option>? fundingInstitutions,
+      @JsonKey(name: "funding_sources")
+          final List<Option>? fundingSources,
+      @JsonKey(name: "gads")
+          final List<Option>? gads,
+      @JsonKey(name: "implementation_modes")
+          final List<Option>? implementationModes,
+      @JsonKey(name: "infrastructure_sectors")
+          final List<Option>? infrastructureSectors,
+      @JsonKey(name: "nafmip_outputs")
+          final List<Option>? nafmipOutputs,
+      @JsonKey(name: "offices")
+          final List<Option>? offices,
+      @JsonKey(name: "operating_units")
+          final List<Option>? operatingUnits,
+      @JsonKey(name: "pdp_chapters")
+          final List<Option>? pdpChapters,
+      @JsonKey(name: "pipol_statuses")
+          final List<Option>? pipolStatuses,
+      @JsonKey(name: "pips_statuses")
+          final List<Option>? pipsStatuses,
+      @JsonKey(name: "preparation_documents")
+          final List<Option>? preparationDocuments,
+      @JsonKey(name: "prerequisites")
+          final List<Option>? prerequisites,
+      @JsonKey(name: "prexcs")
+          final List<Option>? prexcs,
+      @JsonKey(name: "programs")
+          final List<Option>? programs,
+      @JsonKey(name: "project_statuses")
+          final List<Option>? projectStatuses,
+      @JsonKey(name: "provinces")
+          final List<Option>? provinces,
+      @JsonKey(name: "readiness_levels")
+          final List<Option>? readinessLevels,
+      @JsonKey(name: "regions")
+          final List<Option>? regions,
+      @JsonKey(name: "roles")
+          final List<Option>? roles,
+      @JsonKey(name: "sdgs")
+          final List<Option>? sdgs,
+      @JsonKey(name: "spatial_coverages")
+          final List<Option>? spatialCoverages,
+      @JsonKey(name: "types")
+          final List<Option>? types,
+      @JsonKey(name: "typologies")
+          final List<Option>? typologies,
+      @JsonKey(name: "years")
+          final List<Option>? years,
+      @JsonKey(name: "ou_types")
+          final List<Option>? ouTypes,
+      @JsonKey(name: "attachment_types")
+          final List<Option>? attachmentTypes}) = _$_FormOptions;
+
+  factory _FormOptions.fromJson(Map<String, dynamic> json) =
+      _$_FormOptions.fromJson;
+
+  @override
+  @JsonKey(name: "agenda")
+  List<Option>? get agenda;
+  @override
+  @JsonKey(name: "approval_levels")
+  List<Option>? get approvalLevels;
+  @override
+  @JsonKey(name: "bases")
+  List<Option>? get bases;
+  @override
+  @JsonKey(name: "categories")
+  List<Option>? get categories;
+  @override
+  @JsonKey(name: "cip_types")
+  List<Option>? get cipTypes;
+  @override // @JsonKey(name: "commodities") List<Option>? commodities,
+  @JsonKey(name: "fs_statuses")
+  List<Option>? get fsStatuses;
+  @override
+  @JsonKey(name: "funding_institutions")
+  List<Option>? get fundingInstitutions;
+  @override
+  @JsonKey(name: "funding_sources")
+  List<Option>? get fundingSources;
+  @override
+  @JsonKey(name: "gads")
+  List<Option>? get gads;
+  @override
+  @JsonKey(name: "implementation_modes")
+  List<Option>? get implementationModes;
+  @override
+  @JsonKey(name: "infrastructure_sectors")
+  List<Option>? get infrastructureSectors;
+  @override // @JsonKey(name: "locations") List<Option>? locations,
+  @JsonKey(name: "nafmip_outputs")
+  List<Option>? get nafmipOutputs;
+  @override
+  @JsonKey(name: "offices")
+  List<Option>? get offices;
+  @override
+  @JsonKey(name: "operating_units")
+  List<Option>? get operatingUnits;
+  @override
+  @JsonKey(name: "pdp_chapters")
+  List<Option>? get pdpChapters;
+  @override // @JsonKey(name: "pdp_strategies") List<Option>? pdpStrategies,
+  @JsonKey(name: "pipol_statuses")
+  List<Option>? get pipolStatuses;
+  @override
+  @JsonKey(name: "pips_statuses")
+  List<Option>? get pipsStatuses;
+  @override
+  @JsonKey(name: "preparation_documents")
+  List<Option>? get preparationDocuments;
+  @override
+  @JsonKey(name: "prerequisites")
+  List<Option>? get prerequisites;
+  @override
+  @JsonKey(name: "prexcs")
+  List<Option>? get prexcs;
+  @override
+  @JsonKey(name: "programs")
+  List<Option>? get programs;
+  @override
+  @JsonKey(name: "project_statuses")
+  List<Option>? get projectStatuses;
+  @override
+  @JsonKey(name: "provinces")
+  List<Option>? get provinces;
+  @override
+  @JsonKey(name: "readiness_levels")
+  List<Option>? get readinessLevels;
+  @override
+  @JsonKey(name: "regions")
+  List<Option>? get regions;
+  @override
+  @JsonKey(name: "roles")
+  List<Option>? get roles;
+  @override
+  @JsonKey(name: "sdgs")
+  List<Option>? get sdgs;
+  @override
+  @JsonKey(name: "spatial_coverages")
+  List<Option>? get spatialCoverages;
+  @override
+  @JsonKey(name: "types")
+  List<Option>? get types;
+  @override
+  @JsonKey(name: "typologies")
+  List<Option>? get typologies;
+  @override // @JsonKey(name: "vc_segments") List<Option>? vcSegments,
+  @JsonKey(name: "years")
+  List<Option>? get years;
+  @override
+  @JsonKey(name: "ou_types")
+  List<Option>? get ouTypes;
+  @override
+  @JsonKey(name: "attachment_types")
+  List<Option>? get attachmentTypes;
+  @override
+  @JsonKey(ignore: true)
+  _$$_FormOptionsCopyWith<_$_FormOptions> get copyWith =>
+      throw _privateConstructorUsedError;
+}
